pyproject.toml
[project]
name = "OAgent"
version = "0.1.0"
description = "Agentic Chatbot"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "openai-agents>=0.0.6",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.2",
]
